powerpc/paravirt: Improve vcpu_is_preempted
PowerVM Hypervisor dispatches on a whole core basis. In a shared lpar, a
CPU from a core that is ceded or preempted may have a larger latency. In
such a scenario, its preferable to choose a different CPU to run.

If one of the CPUs in the core is active, i.e neither ceded nor
preempted, then consider this CPU as not preempted.

Signed-off-by: Srikar Dronamraju <srikar@linux.vnet.ibm.com>